+++ NINA.Plugin.TargetScheduler/Controls/AcquiredImages/AcquiredImagesManagerViewVM.cs
@@ -200,7 +200,7 @@ namespace NINA.Plugin.TargetScheduler.Controls.AcquiredImages {
                 targets = context.TargetSet.AsNoTracking().Where(t => t.ProjectId == selectedProjectId).ToList();
             }
 
-            targets.ForEach(t => {
+            targets.OrderBy(t => t.Name).ForEach(t => {
                 choices.Add(new KeyValuePair<int, string>(t.Id, t.Name));
             });